(z80asm) refactor strtable and xmalloc

+++ src/common/t/t_strutil.c
@@ -85,7 +85,7 @@ void t_strutil_cstr_strip_compress_escapes(void)
 	for (int i = 0; i < 256; i++)
 	{
 		sprintf(cs, "\\%o \\x%x", i, i);
-		int len = str_compress_escapes(cs);
+		int len = (int)str_compress_escapes(cs);
 		TEST_ASSERT_EQUAL(3, len);
 		TEST_ASSERT_EQUAL((char)i, cs[0]);
 		TEST_ASSERT_EQUAL(' ', cs[1]);
